corrected duplicate files
