+++ src/components/MainPage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import NavBar from './NavBar';
+import NavBar from '../final-project/src/components/NavBar';
 import WelcomeMessage from './WelcomeMessage';
 
 function MainPage() {
